+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Converter for The Netherlands (`nl` and `nl_crops`)
 - Converter for Belgium, Flanders (`be_vlg`)
 - Converter for Slovenia via EuroCrops (`de_si`)
+- Converter for France (`fr`)
 
 ### Changed